move static config into separate folder
