+++ src/config/static_config.hpp
@@ -1,5 +1,5 @@
-#ifndef __VOLUME_RESTIR_STATIC_CONFIG_HPP__
-#define __VOLUME_RESTIR_STATIC_CONFIG_HPP__
+#ifndef __VOLUME_RESTIR_CONFIG_STATIC_CONFIG_HPP__
+#define __VOLUME_RESTIR_CONFIG_STATIC_CONFIG_HPP__
 
 #include <string>
 
@@ -13,4 +13,4 @@ extern const int kWindowHeight;
 }  // namespace static_config
 }  // namespace volume_restir
 
-#endif /* __VOLUME_RESTIR_STATIC_CONFIG_HPP__ */
+#endif /* __VOLUME_RESTIR_CONFIG_STATIC_CONFIG_HPP__ */
